requirements.txt
Pillow
boto3
pyvips